Better retester: make a controller
Some notable changes:
- pwned the fan-out in querying: retest will not need to query in
  parallel and I immediately hit rate limiting when testing
- I removed the `!orgs.Has(org)` which I do not understand why it was
  added
- Added some logging
- Added a `status:failure` into the query to only search failing PRs
  (only ever makes sense to retest failing PRs)
